% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqrf.r
\name{find_max_index}
\alias{find_max_index}
\title{find the closest index}
\usage{
find_max_index(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a value}
}
\value{
the index of the closest value in the vector
}
\description{
This function finds the closest index to a given value in a vector.
}
\examples{
find_max_index(c(1, 2, 3, 4, 5), 3.5)
}
