% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateOutliers.r
\name{generateOutliers}
\alias{generateOutliers}
\title{Adds Outliers}
\usage{
generateOutliers(data, p = 0.05, sd_factor = 5, seed = NULL)
}
\arguments{
\item{data}{data.frame.}

\item{p}{Proportion of outliers to add to data.}

\item{sd_factor}{Each outlier is generated by shifting the original value by a
realization of a normal random variable with \code{sd_factor} times
the original sample standard deviation.}

\item{seed}{An integer seed.}
}
\value{
data with some outliers.
}
\description{
Adds Outliers
}
\examples{
generateOutliers(iris, p = 0.05, sd_factor = 5)
}
