% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqrf.r
\name{get_quantily_value}
\alias{get_quantily_value}
\title{get numberic value from string}
\usage{
get_quantily_value(name)
}
\arguments{
\item{name}{a string}
}
\value{
a numeric value
}
\description{
This function extracts the numeric value from a string.
}
\examples{
get_quantily_value("quantiles = 0.001")
}
