% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqrf.r
\name{get_right_rank}
\alias{get_right_rank}
\title{find the right rank}
\usage{
get_right_rank(response, outMatrix, median_outMatrix, rmse_)
}
\arguments{
\item{response}{a vector of response values}

\item{outMatrix}{a matrix of out values}

\item{median_outMatrix}{a vector of median out values}

\item{rmse_}{a vector of rmse values}
}
\value{
a vector of ranks
}
\description{
This function finds the right rank of a response value in a quantile random forest.
}
