% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method.r
\name{plot.outqrf}
\alias{plot.outqrf}
\title{Plots outqrf}
\usage{
\method{plot}{outqrf}(x, data, ...)
}
\arguments{
\item{x}{An object of class "outqrf".}

\item{data}{Original data frame used for outlier detection. Required.}

\item{...}{other param maybe uesd.}
}
\value{
A ggplot2 object
}
\description{
This function can plot paired boxplot of an "outqrf" object.
It helps us to better observe the relationship between the original and predicted values
}
\examples{
irisWithOutliers <- generateOutliers(iris, seed = 2024)
qrf <- outqrf(irisWithOutliers)
plot(qrf, data = irisWithOutliers)
}
