% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outqrf.r
\name{outqrf}
\alias{outqrf}
\title{find outliers}
\usage{
outqrf(data, quantiles_type = 1000, threshold = 0.025, verbose = 1, ...)
}
\arguments{
\item{data}{a data frame}

\item{quantiles_type}{'1000':seq(from = 0.001, to = 0.999, by = 0.001),'400':seq(0.0025,0.9975,0.0025)}

\item{threshold}{a threshold for outlier detection}

\item{verbose}{a boolean value indicating whether to print verbose output}

\item{...}{additional arguments passed to the ranger function}
}
\value{
a data frame of outliers
}
\description{
This function finds outliers in a dataset using quantile random forests.
}
\examples{
outqrf(iris)
}
